requirements.txt
pycups
protobuf
keyboard
dbus-python
